+++ R/build_log_entry.R
@@ -26,7 +26,7 @@
 #' @param msg.text        Logging message (e. g. error message)
 #' @param call.stack      a call stack created by \code{\link{sys.calls}}
 #' @param dump.file.name  name of the created dump file (leave empty if the \code{\link{tryCatchLog}}
-#'                        argument \code{dump.errors.to.file} is \code{FALSE}
+#'                        argument \code{write.error.dump.file} is \code{FALSE}
 #' @param omit.call.stack.items  the number of stack trace items to ignore (= last x calls) in
 #'                               the passed \code{call.stack} since they are caused by using \code{tryCatchLog}
 #'