+++ tests/testthat/test_last_tryCatchLog_result.R
@@ -7,7 +7,7 @@
 context("last.tryCatchLog.result")
 
 # set up test context
-options("tryCatchLog.dump.errors.to.file" = FALSE)
+options("tryCatchLog.write.error.dump.file" = FALSE)
 options("tryCatchLog.silent.warnings"     = FALSE)
 options("tryCatchLog.silent.messages"     = FALSE)