+++ tests/testthat/test_stacked_tryCatchLog.R
@@ -7,7 +7,7 @@ library(testthat)
 context("stacked tryCatchLog")
 
 # set up test context
-options("tryCatchLog.dump.errors.to.file" = FALSE)
+options("tryCatchLog.write.error.dump.file" = FALSE)
 options("tryCatchLog.silent.warnings"     = FALSE)
 options("tryCatchLog.silent.messages"     = FALSE)